+++ tests/data/ex_prefer_no_split_subscripts.py
@@ -1,0 +1,13 @@
+x["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"] = "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+
+x[x["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"] or "xxxxxxx"] = "xxxxxxxxxxxxxxxxxxxxxxxxx"
+
+# output
+
+x["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"] = (
+    "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+)
+
+x[x["xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"] or "xxxxxxx"] = (
+    "xxxxxxxxxxxxxxxxxxxxxxxxx"
+)