+++ tests/test_format_ex.py
@@ -28,6 +28,13 @@ class TestSimpleFormatEX(BlackBaseTestCase):
             black.Mode(keep_blank_lines_in_brackets=True),
         )
 
+    @patch("black.dump_to_file", dump_to_stderr)
+    def test_prefer_no_split_subscripts(self) -> None:
+        self.check_file(
+            "ex_prefer_no_split_subscripts",
+            black.Mode(prefer_no_split_subscripts=True),
+        )
+
     def check_file(self, filename: str, mode: black.Mode, *, data: bool = True) -> None:
         source, expected = read_data(filename, data=data)
         actual = fs(source, mode=mode)