+++ tests/test_mode_ex.py
@@ -14,6 +14,7 @@ def test_get_cache_key() -> None:
         indentation="\t",
         tab_width=4,
         keep_blank_lines_in_brackets=True,
+        prefer_no_split_subscripts=True,
     ).get_cache_key()
-    expected = "-.1.4.88.1.0.0.1.0.1"
+    expected = "-.1.4.88.1.0.0.1.0.1.1"
     assert result == expected